+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-## 0.4.1 (Unreleased)
+## 0.4.2 (Unreleased)
 
 IMPROVEMENTS:
 
-* Added parameters account and domain_id to cloudstack_resource_instance
+* Added parameters account and domain to cloudstack_resource_instance
+* Added parameters account and domain to cloudstack_resource_network
 
 ## 0.4.0 (Unreleased)